--- src/web/components/ParticipantManagement/ParticipantApprovalForm.scss
@@ -1,34 +1,0 @@
-.participant-approval-form {
-  display: flex;
-  flex-grow: 1;
-  flex-direction: column;
-
-  .highlight {
-    background-color: var(--theme-search-highlight);
-  }
-
-  .search-bar {
-    display: flex;
-    flex-direction: column;
-
-    .site-search-results {
-      flex-grow: 1;
-      overflow: scroll;
-    }
-  }
-
-  .search-input {
-    color: var(--theme-secondary);
-  }
-
-  .site-search-results {
-    padding: 10px;
-
-    & > button {
-      margin: 10px;
-      display: block;
-      font-weight: normal;
-      color: var(--theme-primary);
-    }
-  }
-}
